+++ apps/client/src/app/model/search/fishing-spot-search-result.ts
@@ -1,6 +1,7 @@
 import { BaseSearchResult } from './base-search-result';
+import { LazyFishingSpot } from '../../lazy-data/model/lazy-fishing-spot';
 
 export interface FishingSpotSearchResult extends BaseSearchResult {
   id: number;
-  spot: any;
+  spot: LazyFishingSpot;
 }